+++ fuzzer/src/bin/fuzz-parser-advance.rs
@@ -1,0 +1,10 @@
+use afl::fuzz;
+
+use anes::parser::Parser;
+
+fn main() {
+    fuzz!(|data: &[u8]| {
+        let mut parser = Parser::default();
+        parser.advance(data, false);
+    });
+}