+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+# master
+
+* Add `ReportCursorPosition` sequence
+* Add ANSI escape sequence parser (feature `parser`, disabled by default)
+  * Add `Parser` engine
+  * Add `KeyCode`, `KeyModifiers`, `Mouse`, `MouseButton`, `Sequence` types
+
 # Version 0.1.5
 
 * Add `execute!` macro to execute ANSI escape sequences